src/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
albucore==0.0.24
albumentations==2.0.8
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
annotated-types==0.7.0
bond==4.1.2
builtin-interfaces==1.2.1
cartographer-ros-msgs==2.0.9002
certifi==2025.11.12
charset-normalizer==3.4.4
coloredlogs==15.0.1
composition-interfaces==1.2.1
contourpy==1.3.2
control-msgs==4.8.0
controller==2025.0.0
controller-manager==2.50.0
controller-manager-msgs==2.50.0
cv-bridge==3.2.1
cycler==0.12.1
Cython==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
diagnostic-updater==4.0.4
domain-coordinator==0.10.0
dwb-msgs==1.1.18
dynamixel-sdk==3.8.3
easydict==1.13
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
gazebo-msgs==3.9.0
generate-parameter-library-py==0.4.0
geographic-msgs==1.0.6
geometry-msgs==4.8.0
humanfriendly==10.0
idna==3.11
image-geometry==3.2.1
ImageIO==2.37.2
insightface==0.7.3
interactive-markers==2.3.2
Jinja2==3.1.6
joblib==1.5.2
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.9
laser-geometry==2.4.0
launch==1.0.8
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lazy_loader==0.4
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.3
matplotlib==3.10.7
message-filters==4.3.7
ml_dtypes==0.5.4
mpmath==1.3.0
nav-2d-msgs==1.1.18
nav-msgs==4.8.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nav2-simple-commander==1.0.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
osrf-pycommon==2.1.6
packaging==25.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
prettytable==3.17.0
protobuf==6.33.1
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-qt-binding==1.1.2
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
requests==2.32.5
resource-retriever==3.1.3
rmw-dds-common==1.6.0
robot-localization==3.5.3
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2controlcli==2.50.0
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scripts==3.9.0
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
simsimd==6.5.3
six==1.17.0
slam-toolbox==2.6.10
smclib==4.1.2
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
stringzilla==4.2.3
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
threadpoolctl==3.6.0
tifffile==2025.5.10
topic-monitor==0.20.5
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
trajectory-msgs==4.8.0
triton==3.5.1
turtlebot3-example==2.2.9
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.2.9
turtlesim==1.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics-thop==2.0.18
unique-identifier-msgs==2.2.1
urllib3==2.5.0
vehicle==2025.0.0
vision-msgs==4.1.1
visualization-msgs==4.8.0
wcwidth==0.2.14
webots-ros2==2025.0.0
webots-ros2-crazyflie==2025.0.0
webots-ros2-driver==2025.0.0
webots-ros2-epuck==2025.0.0
webots-ros2-husarion==2025.0.0
webots-ros2-importer==2025.0.0
webots-ros2-mavic==2025.0.0
webots-ros2-msgs==2025.0.0
webots-ros2-tesla==2025.0.0
webots-ros2-tiago==2025.0.0
webots-ros2-turtlebot==2025.0.0
webots-ros2-universal-robot==2025.0.0
xacro==2.0.13
